Don't allow tables with an "sqlkite_" prefix
This might not be strictly necessary, but it seems prudent. For example, access
control CTEs are given a fixed name of "sqlkite_cte_${tableName}". But what if
a project had 2 tables: "users" and "sqlkite_cte_users"? If we don't allow
tables to have an "sqlkite_" prefix, we avoid any ambiguity.