Start work on Table Update
Change how SelectTableAccess.Name is done. This used to be set when the access
was created. But now it's set when the table (project) is loaded. This Name was
essentially a denormalization that we'd persist. Like any denormalization, it
became hard to manage when we do update. Like, what if we change the table name?
We need to make sure to change the SelectTableAccess name...

If we do this on table (project) load, we'll always have the correct name.