Projects.tables map[string]Table -> map[string]*Table
This was a difficult decision. Instances of a project is still immutable, but
the table structure is going to get large, and it's going to be passed around
more and used in more places (e.g. in the Insert and Update HTTP handlers to
validation columns). Sooner or later, we'll want to optimize this, and doing it
now is easier. This does require special care to make sure we don't mutate shared
table references, but, thankfully, there are very explicit places where tables can be updated (namely, project.UpdateTable).