Add tests/sqlkite.super to .gitignore
The plan was that there'd be a working sqlkite.super database that would allow
cloning the repo and running the tests with as little friction or chance of
failure as possible.

However, this creates a ton of noise, since sqlkite.super constantly changes
(like, when we test adding a table, the sqlkite_projects.updated column is
updated).

Need to find an alternative, but until then, the file is being moved out of the
repo.